feat: reminder detail in help command

+++ src/handlers/command/reminder.ts
@@ -8,33 +8,30 @@ import {
 export async function handleReminderCommand(message: Message, args: string) {
   let reply: Message;
 
-  // Try to parse args into an integer
-  const parsedArgs = parseInt(args);
-
   const userCommand = args.split(" ");
   const command = userCommand[0];
 
   switch (command) {
-    case "all":
-      return (reply = await message.reply(
-        await deleteAllReminder(message.from)
-      ));
-    case "list":
-      return (reply = await message.reply(
-        await listAllReminders(message.from)
-      ));
     case "delete":
       const reminderIndex = parseInt(userCommand[1]);
       if (!isNaN(reminderIndex)) {
         return (reply = await message.reply(
           await deleteReminderByIndex(message.from, reminderIndex)
         ));
+      } else if (userCommand[1] === "all") {
+        return (reply = await message.reply(
+          await deleteAllReminder(message.from)
+        ));
       } else {
         return (reply = await message.reply("Invalid reminder index"));
       }
-    default:
+    case "list":
       return (reply = await message.reply(
         await listAllReminders(message.from)
+      ));
+    default:
+      return (reply = await message.reply(
+        "Invalid reminder command. Use `list` to list reminders or `delete <index>` to delete a specific reminder."
       ));
   }
 }
